+++ bsd/net/net_perf.c
@@ -65,7 +65,7 @@ update_bins(net_perf_t *npp, uint64_t bins)
 {
 	bzero(&npp->np_hist_bars, sizeof(npp->np_hist_bars));
 
-	for (int i = 1, j = 0; i <= 64 && j < NET_PERF_BARS; i++) {
+	for (uint8_t i = 1, j = 0; i <= 64 && j < NET_PERF_BARS; i++) {
 		if (bins & 0x1) {
 			npp->np_hist_bars[j] = i;
 			j++;
