+++ bsd/sys/kasl.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013 Apple Inc. All rights reserved.
+ * Copyright (c) 2020 Apple Inc. All rights reserved.
  *
  * @APPLE_OSREFERENCE_LICENSE_HEADER_START@
  *
@@ -37,9 +37,9 @@
 #endif /* BSD_KERNEL_PRIVATE */
 
 extern int
-kern_asl_msg(int level, const char *facility, int num_pairs, ...);
+kern_asl_msg(int level, const char *facility, size_t num_pairs, ...);
 
-extern int escape_str(char *str, int len, int buflen);
+extern int escape_str(char *str, size_t len, size_t buflen);
 extern void fpxlog_init(void);
 extern void fpxlog(int, uint32_t, uint32_t, uint32_t);
