+++ bsd/sys/sys_domain.h
@@ -56,7 +56,7 @@ struct sockaddr_sys {
 #ifdef PRIVATE
 struct  xsystmgen {
 	u_int32_t       xg_len; /* length of this structure */
-	u_int32_t       xg_count;       /* number of PCBs at this time */
+	u_int64_t       xg_count;       /* number of PCBs at this time */
 	u_int64_t       xg_gen; /* generation count at this time */
 	u_int64_t       xg_sogen;       /* current socket generation count */
 };
