+++ security/mac_system.c
@@ -207,7 +207,7 @@ mac_system_check_swapoff(kauth_cred_t cred, struct vnode *vp)
 
 int
 mac_system_check_sysctlbyname(kauth_cred_t cred, const char *namestring, int *name,
-    u_int namelen, user_addr_t oldctl, size_t oldlen,
+    size_t namelen, user_addr_t oldctl, size_t oldlen,
     user_addr_t newctl, size_t newlen)
 {
 	int error;
